Set local commonmark immediately after loading it
Prior to this commit, a local variable named `commonmark` was being set
to `window.commonmark` in the `renderPage()` function. This local
variable was then used to render Markdown. The problem with this
approach was that if the user code did not call `mdme.renderPage()` but
instead called `mdme.render()` directly, then the local variable
`commonmark` was never being set, so it was left as `undefined`. Using
this variable led to error.

This commit sets the local variable named `commonmark` as soon as it
loads to ensure that it is available for use by any function that depens
on it.